merge

+++ src/main/java/istad/co/exstadbackendapi/features/user/UserRepository.java
@@ -1,12 +1,11 @@
 package istad.co.exstadbackendapi.features.user;
 
-import aj.org.objectweb.asm.commons.Remapper;
 import istad.co.exstadbackendapi.domain.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
     Optional<User> findByUsername(String username);
     Optional<User> findByUuid(String uuid);
 }
